buyhatke/requirements.txt
asgiref==3.8.1
bson==0.5.10
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.1.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
idna==3.7
numpy==1.26.4
pandas==2.2.2
PyJWT==2.8.0
pymodm==0.4.3
pymongo==4.6.3
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.2.4
tzdata==2024.1
urllib3==2.2.1
